+++ src/ca/squadcar/games/editor/IDrawableElement.java
@@ -6,5 +6,5 @@ public interface IDrawableElement {
 
 	public void draw(Graphics gfx, float zoomFactor);
 	
-	public boolean hitTest(final int x, final int y, final float zoomFactor);
+	public boolean hitTest(final float x, final float y, final float zoomFactor);
 }
